% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertAddin.R
\name{insertCompAssAddin}
\alias{insertCompAssAddin}
\title{Insert \%<>\% (magrittr compound assignment pipe-operator).}
\usage{
insertCompAssAddin()
}
\description{
Call this function as an addin to insert ` \%<>\% ` at the cursor position.
}
\seealso{
[magrittr::`%<>%`]
}
