% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertAddin.R
\name{insertExpAddin}
\alias{insertExpAddin}
\title{Insert \%$\% (magrittr exposition pipe-operator).}
\usage{
insertExpAddin()
}
\description{
Call this function as an addin to insert \code{ \%$\% } at the cursor position.
}
