% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertAddin.R
\name{insertGetAttrAddin}
\alias{insertGetAttrAddin}
\title{Insert \%@\% (rlang attribute getter pipe-operator).}
\usage{
insertGetAttrAddin()
}
\description{
Call this function as an addin to insert ` \%@\% ` at the cursor position.
}
\seealso{
[rlang::`%@%`]
}
