% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insertAddin.R
\name{insertToppAddin}
\alias{insertToppAddin}
\title{Insert \%T>\% (magrittr tee operator).}
\usage{
insertToppAddin()
}
\description{
Call this function as an addin to insert ` \%T>\% ` at the cursor position.
}
\seealso{
[magrittr::`%T>%`]
}
